VgaCanvas: Reset attributes on each EOL
Always emit a "reset attributes" on EOL in VgaCanvas_serialize() as some
terminals will misbehave when a background color is active on newline.
Add a flag to select whether to "chop" lines or not, only set by main
when the output will use the terminal's default colors.

Adapt AnsiColorWriter to work well with this change. Overall, this
should fix any display issues on line endings on all (most?) terminals.

Also add command-line flags to enable all the new features.