WinConsoleWriter: write using legacy Console API
Add a writer for older Windows versions not supporting ANSI sequences at
all. Force using it on Windows <10 when output goes to the console. For
Windows 10 and newer, configure the AnsiColorWriter to use "exact
colors" (256 color mode) unless the user forces simple ANSI with -a,
because we know Windows Terminal supports this.