TiColorWriter: Write to stream instead of stdout
Use tputs to have terminfo/curses write to the stream provided. This
eliminates the need for reliable flushing and makes it possible to use
this writer for output to files and for other output formats than UTF8.

This isn't threadsafe (and can't be because of the terminfo design). It
will need some lock for single instance if threading is added.