TiColorWriter: Use terminfo capabilities directly
Terminfo provides macros for directly accessing the capabilities, so
there is no need to use tigetnum()/tigetstr() at all.

This also removes the fallback to query capabilities by their termcap
name, which is pointless anyways (it *should* just access the same
capabilities) and lead to a strange bug at least with the terminfo entry
for xterm-256color in FreeBSD, which for some reason returns "something"
for "mb" which unfortunately isn't the capability for "blink".