SaucePrinter: Re-format comments nicely
It seems to be common practice with SAUCE comments to wrap lines in the
middle of a word. Try to detect this and apply some real word-wrapping
using the following rules:

* Empty lines and lines containing "drawing" characters (>= 0xB0) are
  printed as is
* Words are joined across lines when the last character of the first
  line and the first character of the second line are non-space
* Words are filled into new formatted lines, preserving original
  whitespace between them except at the end of a line
* When a line ends with more than 2 space characters, an intentional
  line break (paragraph) is assumed and a new formatted line is started

For this to work, preserve one trailing space in a "short" commend line
in the Sauce parser.

Add a new flag to Config (-W: nowrap) to disable this logic and print
the comment lines as is.