DosReader: Simplify and add maximum size
There's no need to resume reading after "DOS EOF" any more, so this code
can be dropped, and we can just use the normal SS_EOF stream status in
any case.

Instead, add a maximum size to set this to the start position of SAUCE
metadata. When not ignoring EOF, reading will stop at latest there (or
earlier if a "DOS EOF" character is encountered).