showansi: Use POSIX/C locale
This hopefully fixes byte-wise string processing for *most* POSIX
compliant shells, tested with Schily's "pbosh". It might still be
possible to find a shell that refuses to process any characters >0x7f in
the C locale, but there's really nothing we can do about *that*.

Invoke the terminal with the original locale.